\name{iptw-package}
\alias{iptw-package}
\docType{package}
\title{
Inverse Probability of Treatment Weighting
}
\description{
This package calculates various measures of association and helps understand and visualize the link between causal models such as Marginal Structural Models and Standerdized measures.}
\details{
\tabular{ll}{
Package: \tab iptw\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-25\cr
License: \tab GPL 2 or newer\cr
}

}
\author{
Ehsan Karim <ehsan@alumni.ubc.ca>

Maintainer: Ehsan Karim <ehsan@alumni.ubc.ca>
}

\keyword{Causal}

